feat(pagination): create adapter for Laravel's paginator
This commit introduces an abstraction layer over Laravel's
`LengthAwarePaginator` to standardize the structure of pagination
data in the API, especially the link format.

The main goal is to decouple the rest of the application from
Laravel's specific paginator implementation. This is achieved
through a custom `PaginatorInterface`, ensuring the pagination
structure in the API response is always consistent.

- The `PaginatorLengthAwarePaginator` class acts as an Adapter.
- The link format (`first`, `last`, `prev`, `next`) is now standardized.