test(dto): add unit tests for InputDtoInterface contract
This commit adds unit tests to validate the contract established by
the `InputDtoInterface`.

Since an interface cannot be tested directly, a test stub
(`CreatePostInputDtoStub`) was created to serve as a concrete
implementation. The test ensures that the `toArray()` method correctly
converts the DTO's data into an array.

A Data Provider is used to validate multiple scenarios, including
nullable fields.