refactor(repository): allow models to define their own query pipelines
This refactor introduces a new architecture that allows Eloquent models
to define their own sets of Query Pipelines to be applied.

The goal is to make the system more extensible and to give models the
responsibility for their own query logic, decoupling the repository
from the definition of which filters to apply.

- A new `ModelCrudInterface` was created with the static methods
  `pipelinesFindAll()` and `pipelinesFindOne()`.
- The `User` model now implements this interface.
- The `EloquentUserRepository` has been refactored to consume these
  methods from the model, instead of using an empty pipeline list.