refactor(core): decouple `setResultResource` from HTTP request
This refactor improves the robustness of the `ResponseBuilder` by
removing the dependency on the `Request` object when resolving
a `JsonResource`.

The call to `$resource->toArray($request)` has been replaced with
`$resource->jsonSerialize()`.

This makes the builder more versatile and safer to use in different
contexts beyond an HTTP request, such as in Jobs, Commands, or
tests, without the risk of capturing the incorrect request.